Added the image in the chat
